Calculate the start column always
YouCompleteMe can use omnifunc, but it calls that differently from Vim.
Before making the second call, Vim sets the cursor position returned by
the first call. YouCompleteMe doesn't do that. As a result, getcurpos()
on the second call can be different. To amend this, always calculate the
start column and use that instead of getcurpos().